Convert Collection to TS #355
